feat(run): moved train and compute to pycmtensor.run for cleaniness
