+++ pycmtensor/models/__init__.py
@@ -1,3 +1,2 @@
-from pycmtensor.models.basic import compute, train
 from pycmtensor.models.MNL import MNL
 from pycmtensor.models.TasteNet import TasteNet
